Ajouter compagnonage know takes advantage of the compagnonage-list livewire component

+++ app/resources/views/livewire/compagnonage-list.blade.php
@@ -21,15 +21,25 @@
                         <th scope="row">{{ $compagnonage->id }}</th>
                         <td>{{ $compagnonage->comp_libcourt }}</td>
                         <td>{{ $compagnonage->comp_liblong }}</td>
-                        <td><a href="{{ route('compagnonages.show', $compagnonage->id) }}" class="btn btn-primary btn-sm">Consulter</a></td>
-                        <td><a href="{{ route('compagnonages.edit', $compagnonage->id) }}" class="btn btn-info btn-sm">Editer</a></td>
-                        @can('compagnonages.destroy')
-                        <td>
-                            {!! Form::open(['method' => 'DELETE','route' => ['compagnonages.destroy', $compagnonage->id],'style'=>'display:inline']) !!}
-                            {!! Form::submit('Supprimer', ['class' => 'btn btn-danger btn-sm']) !!}
-                            {!! Form::close() !!}
-                        </td>
-                        @endcan
+                        @if ($mode == "gestion")
+                            <td><a href="{{ route('compagnonages.show', $compagnonage->id) }}" class="btn btn-primary btn-sm">Consulter</a></td>
+                            <td><a href="{{ route('compagnonages.edit', $compagnonage->id) }}" class="btn btn-info btn-sm">Editer</a></td>
+                            @can('compagnonages.destroy')
+                            <td>
+                                {!! Form::open(['method' => 'DELETE','route' => ['compagnonages.destroy', $compagnonage->id],'style'=>'display:inline']) !!}
+                                {!! Form::submit('Supprimer', ['class' => 'btn btn-danger btn-sm']) !!}
+                                {!! Form::close() !!}
+                            </td>
+                            @endcan
+                        @elseif ($mode == "selection")
+                            <td>
+                                {!! Form::open(['method' => 'POST','route' => ['fonctions.ajoutercompagnonage', $fonction->id] ]) !!}
+                                <input type='hidden' id='compagnonage_id' name='compagnonage_id' value='{{ $compagnonage->id }}'>
+                                <button type="submit" class="btn btn-primary btn-sm">Ajouter</a></td>
+                                {!! Form::close() !!}
+                            <td></td>
+                            <td></td>
+                        @endif
                     </tr>
                 @endforeach
             </tbody>
